Code cleaned up
