+++ src/components/video-stream.tsx
@@ -1,19 +1,45 @@
-import { useRef, useEffect, type ReactNode } from "react";
+import type { StreamClient } from "@viamrobotics/sdk";
+import { useRef, useEffect, type ReactNode, useState } from "react";
 
 export interface VideoStreamProps {
-  stream?: MediaStream;
+  streamClient: StreamClient;
+  cameraName: string;
   children?: ReactNode;
 }
 
 export const VideoStream = (props: VideoStreamProps): JSX.Element => {
-  const { stream, children } = props;
+  const { streamClient, cameraName, children } = props;
   const videoRef = useRef<HTMLVideoElement>(null);
+  const okToConnectRef = useRef(true);
+  const [stream, setStream] = useState<MediaStream | undefined>(undefined);
+
+  if (videoRef.current && stream?.active) {
+    videoRef.current.srcObject = stream;
+  }
 
   useEffect(() => {
-    if (videoRef.current && stream) {
-      videoRef.current.srcObject = stream;
+    if (streamClient && okToConnectRef.current) {
+      okToConnectRef.current = false;
+      streamClient
+        .getStream(cameraName)
+        .then((stream) => {
+          if (videoRef.current) {
+            setStream(stream);
+          }
+        })
+        .catch((error) => {
+          console.warn(`Unable to connect to camera ${cameraName}`, error);
+        });
     }
-  }, [stream]);
+    return () => {
+      if (stream && stream?.active) {
+        streamClient.remove("camera").catch((error: unknown) => {
+          console.warn(`Unable to disconnect to camera ${cameraName}`, error);
+          okToConnectRef.current = true;
+        });
+      }
+    };
+  }, []);
 
   return (
     <div className="relative inline-flex p-4">
